requirements.txt
Django==4.1.2
djangorestframework==3.14.0
drf-yasg==1.21.4
django-cors-headers==3.13.0
django-environ==0.9.0
psycopg2-binary==2.9.4
PyJWT==2.5.0
